Client: improve "announcements"
- move announcement within layout
- integrate "poverty accept/reject" buttons and announcement into
"announcement area"
