+++ server-resources/src/main/resources/web/index.html
@@ -12,7 +12,7 @@
 <div id="announcements"></div>
 
 <div id="trick"></div>
-<div id="povertybuttons"></div>
+
 <p></p>
 <div id="bidbuttons"></div>
 <p></p>
